+++ union_sync_impl/src/main/java/com/example/union_sync_impl/entity/ReserveUpdate.kt
@@ -1,11 +1,12 @@
 package com.example.union_sync_impl.entity
 
 import androidx.room.Entity
+import java.math.BigDecimal
 
 @Entity
 data class ReserveUpdate(
     val id: String,
-    val count: Long,
+    val count: BigDecimal,
     val locationId: String?,
     val updateDate: Long?,
     val userUpdated: String?
